Fixed the release version of the project Fixed the structure of the project (added folders and such)
Added phyxels.
Added a phyxel grid.
Added a ChangedPhyxel class that contains the phyxels that contain a certain force that is acting on them.
	- Phyxels are added to this class when a force acts on it, when there is no force acting on it, the phyxel is removed from the list


Took out the octree, it was proving difficult to use since the surfels could be in an intersection of the first child octreenodes and therefore be in the parent octree node